api/requirements.txt
fastapi
fastapi[standard]
pandas
httpx
sqlalchemy
psycopg
psycopg2
python-dotenv